+++ console/src/features/MetadataAPI/hooks/useMetadataMigration.ts
@@ -1,5 +1,5 @@
-import returnMigrateUrl from '@/components/Services/Data/Common/getMigrateUrl';
 import { CLI_CONSOLE_MODE, SERVER_CONSOLE_MODE } from '@/constants';
+import Endpoints from '@/Endpoints';
 import { useMigrationMode } from '@/hooks';
 import { Api } from '@/hooks/apiUtils';
 import { RunSQLResponse } from '@/hooks/types';
@@ -7,7 +7,8 @@ import { useConsoleConfig } from '@/hooks/useEnvVars';
 import { useAppSelector } from '@/store';
 import { useMutation, UseMutationOptions, useQueryClient } from 'react-query';
 import sanitize from 'sanitize-filename';
-import { allowedMetadataTypes } from '../types';
+import { allowedMetadataTypes, MetadataResponse } from '../types';
+import { returnMigrateUrl } from './utils';
 
 const maxAllowedLength = 255;
 const unixEpochLength = 14;
@@ -23,7 +24,8 @@ export function useMetadataMigration(
   mutationOptions?: Omit<
     UseMutationOptions<Record<string, any>, Error, TMigration>,
     'mutationFn'
-  >
+  >,
+  overrideCliMode?: boolean
 ) {
   const { mode } = useConsoleConfig();
   const headers = useAppSelector(state => state.tables.dataHeaders);
@@ -35,11 +37,15 @@ export function useMetadataMigration(
       try {
         const { source, query, migrationName } = props;
 
-        const migrateUrl = returnMigrateUrl(migrationMode ?? false, [query]);
+        const migrateUrl = returnMigrateUrl(
+          migrationMode ?? false,
+          [query],
+          overrideCliMode
+        );
 
         let body = {};
 
-        if (mode === SERVER_CONSOLE_MODE) {
+        if (mode === SERVER_CONSOLE_MODE || overrideCliMode) {
           body = query;
         } else {
           body = {
@@ -69,6 +75,16 @@ export function useMetadataMigration(
       onSuccess: (data, variables, ctx) => {
         if (mode === CLI_CONSOLE_MODE) {
           queryClient.refetchQueries('migrationMode', { active: true });
+          queryClient.fetchQuery({
+            queryKey: 'cliExport',
+            queryFn: () => {
+              const cliMetadataExportUrl = `${Endpoints.hasuractlMetadata}?export=true`;
+              return Api.get<MetadataResponse>({
+                headers,
+                url: cliMetadataExportUrl,
+              });
+            },
+          });
         }
 
         queryClient.refetchQueries(['metadata'], { active: true });