+++ console/src/features/RelationshipsTable/components/SourceColumnCell.tsx
@@ -4,15 +4,34 @@ import { RelationshipSourceType } from '../types';
 
 const getRelationIcon = (type: RelationshipSourceType) => {
   if (type === 'local_object')
-    return <FaTable className="fill-current text-muted text-sm mr-2" />;
+    return (
+      <FaTable className="fill-current text-muted text-sm mr-2" title="Table" />
+    );
   if (type === 'local_array')
-    return <FaTable className="fill-current text-muted text-sm mr-2" />;
+    return (
+      <FaTable className="fill-current text-muted text-sm mr-2" title="Table" />
+    );
   if (type === 'to_source')
-    return <FaDatabase className="fill-current text-muted text-sm mr-2" />;
+    return (
+      <FaDatabase
+        className="fill-current text-muted text-sm mr-2"
+        title="Database"
+      />
+    );
   if (type === 'to_remote_schema')
-    return <FaPlug className="fill-current text-muted text-sm mr-2" />;
+    return (
+      <FaPlug
+        className="fill-current text-muted text-sm mr-2"
+        title="Remote schema"
+      />
+    );
   if (type === 'remote_schema_legacy')
-    return <FaPlug className="fill-current text-muted text-sm mr-2" />;
+    return (
+      <FaPlug
+        className="fill-current text-muted text-sm mr-2"
+        title="Remote schema"
+      />
+    );
 };
 
 type SourceColumnCellType = {